requirements.txt
pyyaml>=6.0
requests>=2.31.0
arxiv>=2.1.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
dashscope>=1.14.0
lxml>=4.9.0
schedule>=1.2.0
